refactor: Move DefinitionsStrategy from common to extension
